[REFACTOR] 마이페이지 조회 API 리팩토링
