+++ src/main/java/com/smeme/server/service/DiaryService.java
@@ -22,9 +22,9 @@ import com.smeme.server.model.Diary;
 import com.smeme.server.model.Member;
 import com.smeme.server.model.badge.Badge;
 import com.smeme.server.model.topic.Topic;
-import com.smeme.server.repository.BadgeRepository;
+import com.smeme.server.repository.badge.BadgeRepository;
 import com.smeme.server.repository.correction.CorrectionRepository;
-import com.smeme.server.repository.MemberBadgeRepository;
+import com.smeme.server.repository.badge.MemberBadgeRepository;
 import com.smeme.server.repository.diary.DiaryRepository;
 import com.smeme.server.repository.MemberRepository;
 import com.smeme.server.repository.topic.TopicRepository;
